feat: rename js/sentry to js/telemetry
