+++ js/telemetry.ts
